Add a dev helper script.
Break the event trigger handler into its own function so it can be
run outside the ssh listener, and then use this in the dev script.
This makes it possible to test develop the bot without needing to
actually poll a gerrit instance.

Clearly, further testing should be performed on a real instance,
but this vastly simplifies the edit-debug cycle.