Undo handling of different committer/uploader.
It appears that while gerrit presents user data to you in three
places during the upload event, these are always the same, even
when the committer data is different to the pushing account.
This code is therefore redundant, and may not be correct even if
this functionality did appear at some point.