Introduce a subscription model for event handlers.
Having each handler subscribe to events via a decorator makes it
clearer from the handlers which each one is used for. Move the
handlers and routing functionality into their own files, now that
it is convenient to do so.

This also makes it very easy to add more handlers, and further provides
separation between each of the parts of the system.

Change-Id: If49d81953dbac9a44110034db711634fbd47b3eb